Agentic_MultiModal_RAG/pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "graphviz>=0.21",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langsmith>=0.7.14",
    "pip>=26.0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
